requirements.txt
# Python
python-decouple
Django>=4.2,<5.0
gunicorn>=21.2
psycopg2-binary>=2.9
whitenoise>=6.6
django-anymail[mailgun]
django-cors-headers
Pillow>=10.0.0
google-generativeai>=0.3.0
